fix: minor tweaks
